Massive user code refactor
