+++ keyboards/rgbkb/zygomorph/keymaps/xulkal/keymap.c
@@ -67,21 +67,3 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
   ),
 #endif
 };
-
-#ifdef ENCODER_ENABLE
-void encoder_update_user(uint8_t index, bool clockwise) {
-  if (index == 0) { /* First encoder */
-    if (clockwise) {
-      tap_code(KC_PGDN);
-    } else {
-      tap_code(KC_PGUP);
-    }
-  } else if (index == 1) { /* Second encoder from slave */
-    if (clockwise) {
-      tap_code(KC_UP);
-    } else {
-      tap_code(KC_DOWN);
-    }
-  }
-}
-#endif
