+++ users/xulkal/custom_encoder.h
@@ -1,0 +1,2 @@
+#pragma once
+#include "quantum.h"
