+++ users/xulkal/rules.mk
@@ -1,10 +1,13 @@
 SRC += xulkal.c \
-       process_records.c
+    process_records.c \
+    custom_tap_dance.c \
+    custom_encoder.c \
+    custom_oled.c \
+    timer_utils.c
 
 # Some usual defaults
 MOUSEKEY_ENABLE         = no    # Mouse keys (+4700)
 EXTRAKEY_ENABLE         = yes   # Audio control and System control (+450)
-TAP_DANCE_ENABLE        = no   # Enable the tap dance feature. (+1100)
 
 ifneq ($(strip $(DISABLE_LTO)), yes)
   EXTRAFLAGS += -flto
