+++ users/xulkal/timer_utils.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "timer.h"
+#include <stdbool.h>
+
+bool timer_expired(uint16_t last);
+bool timer_expired32(uint32_t last);
