+++ users/xulkal/xulkal.h
@@ -2,3 +2,6 @@
 
 #include "process_records.h"
 #include "layouts.h"
+#include "timer_utils.h"
+#include "custom_keycodes.h"
+#include "custom_tap_dance.h"
